Removed numpy.float (deprecated) and coverted to float

+++ rdkit/ML/Composite/AdjustComposite.py
@@ -45,7 +45,7 @@ def BalanceComposite(model, set1, set2, weight, targetSize, names1=None, names2=
   res.quantizationRequirements = []
 
   startSize = len(model)
-  scores = numpy.zeros(startSize, numpy.float)
+  scores = numpy.zeros(startSize, float)
   actQuantBounds = model.GetActivityQuantBounds()
   if names1 is not None:
     model.SetInputOrder(names1)
